+++ lib/src/size_extension.dart
@@ -20,9 +20,12 @@ extension SizeExtension on num {
   ///smart size :  it check your value - if it is bigger than your value it will set your value
   ///for example, you have set 16.sm() , if for your screen 16.sp() is bigger than 16 , then it will set 16 not 16.sp()
   ///I think that it is good for save size balance on big sizes of screen
+  double get spMin => min(toDouble(), sp);
+
+  @Deprecated('use spMin instead')
   double get sm => min(toDouble(), sp);
 
-  double get smx => max(toDouble(), sp);
+  double get spMax => max(toDouble(), sp);
 
   ///屏幕宽度的倍数
   ///Multiple of screen width